backend/requirements.txt
Django
djangorestframework
uritemplate
pyyaml
gunicorn
django-on-heroku